refactor!: stop leaking chumsky prql_parser
